projects/hash_tag_gen/requirements.txt
streamlit
openai-whisper
ffmpeg-python
opencv-python-headless
scikit-image
google-generativeai
numpy
noisereduce
soundfile

